Create CODE_DAY_CALCULATOR
Description:
A simple Python program that calculates the difference between two dates. It provides the following outputs:

The number of days (as a float)
The day of the week for the initial date (as a string)
The number of weeks (as an integer)
The number of years (as an integer)
Features:
Input for the start date.
Input for the end date.
Output includes:
Total number of days between the two dates.
The day of the week corresponding to the start date.
The total number of weeks between the dates.
The total number of years between the dates.